refactor: 단일 Message 관련 컴포넌트들을 폴더로 묶기
