+++ src/app/chat/[chatId]/_components/OtherMessageGroup.tsx
@@ -1,5 +1,5 @@
-import { MessageHeader } from "./MessageHeader"
-import { MessageContent } from "./MessageContent"
+import { MessageHeader } from "./Message/MessageHeader"
+import { MessageContent } from "./Message/MessageContent"
 import { Message, UserInfo } from "./types"
 
 export interface OtherMessageGroupProps {
